+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - **bugsnag:** support stack trace
+- **bugsnag:** context and error name
 
 ### Changed